--- src/features/summoner/enum.ts
@@ -1,16 +1,0 @@
-export enum Chef {
-  SUMMONER,
-  SUMMONER_V2,
-  MINICHEF,
-}
-
-export enum Rewarder {
-  SIMPLE,
-  COMPLEX,
-  ALCX,
-}
-
-export enum PairType {
-  SWAP,
-  KASHI,
-}
