+++ src/hooks/index.ts
@@ -6,3 +6,6 @@ export * from './useContract'
 export { default as useFuse } from './useFuse'
 export { default as useSortableData } from './useSortableData'
 export { default as useUSDCPrice } from './useUSDCPrice'
+export * from './usePrice'
+export * from './useFarmRewards'
+export * from './useSummonerInfo'
