+++ src/modals/LuxorStatsModal/index.tsx
@@ -7,7 +7,7 @@ import HeadlessUiModal from 'components/Modal/HeadlessUIModal'
 import Typography from 'components/Typography'
 import ExternalLink from 'components/ExternalLink'
 import { ExternalLink as LinkIcon } from 'react-feather'
-import { useTokenInfo } from 'features/summoner/hooks'
+import { useTokenInfo } from 'hooks/useTokenInfo'
 import { useWrappedLumensContract, useLuxorContract } from 'hooks'
 import { formatNumberScale } from 'functions'
 import { LUX_ADDRESS, WLUM_ADDRESS } from 'constants/addresses'
