+++ src/modals/TokensStatsModal/index.tsx
@@ -6,7 +6,7 @@ import styled from 'styled-components'
 import Typography from 'components/Typography'
 import ExternalLink from 'components/ExternalLink'
 import { ExternalLink as LinkIcon } from 'react-feather'
-import { useTokenInfo } from 'features/summoner/hooks'
+import { useTokenInfo } from 'hooks/useTokenInfo'
 import { useSeanceContract, useSoulContract } from 'hooks'
 import { formatNumberScale } from 'functions'
 import { SOUL_ADDRESS, SEANCE_ADDRESS } from 'constants/addresses'
