+++ src/pages/mines/index.tsx
@@ -6,7 +6,6 @@ import Search from 'components/Search'
 import MineList from 'features/mines/MineList'
 import Menu from 'features/mines/components/MineMenu'
 // import Header from 'features/mines/components/Header'
-import { useFarms, useSummonerInfo } from 'features/summoner/hooks'
 // import { classNames } from 'functions/styling'
 import useFuse from 'hooks/useFuse'
 import { useActiveWeb3React } from 'services/web3'
@@ -18,10 +17,9 @@ import { POOLS } from 'constants/farms'
 // import { Button } from 'components/Button'
 // import { formatNumberScale } from 'functions'
 // import { addTransaction } from 'state/transactions/actions'
-// import useSummoner from 'features/summoner/useSummoner'
 // import { getAddress } from '@ethersproject/address'
 // import { useTVL } from 'hooks/useV2Pairs'
-import { usePrice } from 'hooks/usePrice'
+import { usePrice, useFarms, useSummonerInfo } from 'hooks'
 // import { SEANCE_ADDRESS, WNATIVE } from 'constants/addresses'
 import { TridentBody, TridentHeader } from 'layouts/Trident'
 import Typography from 'components/Typography'
