+++ src/pages/vaults/index.tsx
@@ -13,15 +13,13 @@ import { useLingui } from '@lingui/react'
 import DoubleGlowShadowV2 from '../../components/DoubleGlowShadowV2'
 import { SOUL_ADDRESS, AVERAGE_BLOCK_TIME, WNATIVE } from '../../constants'
 import { VAULTS } from '../../constants/vaults'
-// import SoulSwapLogo from '../../components/SoulSwapLogo'
 import { PriceContext } from '../../contexts/priceContext'
 import { useTVL } from '../../hooks/useV2Pairs'
-// import { getAddress } from '@ethersproject/address'
 import VaultList from '../../features/vault/VaultList'
-// import useSummoner from '../../features/summoner/useSummoner'
-import { useFarms, usePositions, useSoulSummonerInfo } from '../../features/summoner/hooks'
+import { useFarms } from 'hooks/useFarmRewards'
+import { usePositions } from 'hooks/usePositions'
+import { useSoulSummonerInfo } from 'hooks/useSummonerInfo'
 import { useActiveWeb3React } from 'services/web3'
-// import useSoulSummoner from '../../features/farm/hooks/useSoulSummoner'
 
 export default function Vault(): JSX.Element {
   const { i18n } = useLingui()
